+++ src/app/navigation/navigation.module.ts
@@ -74,6 +74,7 @@ import { CalendarsStyleComponent } from '../styles/calendars-style/calendars-sty
 import { NgxColorsModule } from 'ngx-colors';
 import { SbAsterChartStyleComponent } from '../styles/sb-aster-chart-style/sb-aster-chart-style.component';
 import { AutofillDirective } from '../autofill.directive';
+import { ShortcutButtonStyleComponent } from '../styles/shortcut-button-style/shortcut-button-style.component';
 PlotlyModule.plotlyjs = PlotlyJS;
 
 const routes: Routes = [
@@ -180,7 +181,8 @@ const routes: Routes = [
         ModalStyleComponent,
         CalendarsStyleComponent,
         SbAsterChartStyleComponent,
-        AutofillDirective
+        AutofillDirective,
+        ShortcutButtonStyleComponent
     ]
 })
 export class NavigationPageModule {