+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 3.2.0
  - add [capacitor-native-settings](https://github.com/RaphaelWoude/capacitor-native-settings) v5.0.1
+ - add style `shortcutButton` - open native window
 
 # 3.1.10
  - load `toggle switch` when it is enabled for a input from type `checkbox`